add migration and data types
