+++ packages/model/src/user.ts
@@ -1,4 +1,4 @@
-import { FetchOptions, ListResponse } from './common';
+import { FetchOptions, ListResponse, OrcidWork } from './common';
 import { InterestGroupMembership } from './interest-group';
 import { LabResponse } from './lab';
 import { TeamRole } from './team';
@@ -40,66 +40,6 @@ export interface Invitee {
   institution?: string;
   jobTitle?: string;
   lastName: string;
-}
-
-export const orcidWorkType = [
-  'ANNOTATION',
-  'ARTISTIC_PERFORMANCE',
-  'BOOK_CHAPTER',
-  'BOOK_REVIEW',
-  'BOOK',
-  'CONFERENCE_ABSTRACT',
-  'CONFERENCE_PAPER',
-  'CONFERENCE_POSTER',
-  'DATA_SET',
-  'DICTIONARY_ENTRY',
-  'DISCLOSURE',
-  'DISSERTATION',
-  'EDITED_BOOK',
-  'ENCYCLOPEDIA_ENTRY',
-  'INVENTION',
-  'JOURNAL_ARTICLE',
-  'JOURNAL_ISSUE',
-  'LECTURE_SPEECH',
-  'LICENSE',
-  'MAGAZINE_ARTICLE',
-  'MANUAL',
-  'NEWSLETTER_ARTICLE',
-  'NEWSPAPER_ARTICLE',
-  'ONLINE_RESOURCE',
-  'OTHER',
-  'PATENT',
-  'PHYSICAL_OBJECT',
-  'PREPRINT',
-  'REGISTERED_COPYRIGHT',
-  'REPORT',
-  'RESEARCH_TECHNIQUE',
-  'RESEARCH_TOOL',
-  'SOFTWARE',
-  'SPIN_OFF_COMPANY',
-  'STANDARDS_AND_POLICY',
-  'SUPERVISED_STUDENT_PUBLICATION',
-  'TECHNICAL_STANDARD',
-  'TEST',
-  'TRADEMARK',
-  'TRANSLATION',
-  'WEBSITE',
-  'WORKING_PAPER',
-  'UNDEFINED',
-] as const;
-export type OrcidWorkType = (typeof orcidWorkType)[number];
-
-export interface OrcidWork {
-  id: string;
-  doi?: string;
-  title?: string;
-  type: OrcidWorkType;
-  publicationDate: {
-    year?: string;
-    month?: string;
-    day?: string;
-  };
-  lastModifiedDate: string;
 }
 
 export interface UserTeam {
